requirements.txt
spacy-streamlit>=1.0.0a0
streamlit
pandas
unidecode